requirements.txt
aiofiles==23.2.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.120.0
ffmpy==0.6.4
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.9.0
gradio==4.44.1
gradio_client==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.23.2
humanfriendly==10.0
idna==3.11
importlib_resources==6.5.2
ipykernel==7.0.1
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.3
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
onnx==1.16.2
onnxruntime-gpu==1.18.0
onnxslim==0.1.31
opencv-python==4.9.0.80
opencv-python-headless==4.11.0.86
orjson==3.11.4
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==10.4.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
protobuf==6.33.0
psutil==5.9.8
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycocotools==2.0.7
pydantic==2.12.3
pydantic_core==2.41.4
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.1
pyzmq==27.1.0
requests==2.32.5
rich==14.2.0
ruff==0.14.2
safetensors==0.4.3
scipy==1.13.0
seaborn==0.13.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.48.0
supervision==0.22.0
sympy==1.14.0
timm==1.0.14
tomlkit==0.12.0
torch==2.2.2+cu121
torchvision==0.17.2+cu121
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.176
ultralytics-thop==2.0.17
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
websockets==12.0
